add voc
